RDKE-143 Update bblayers.conf.sample to include all layers
In order to support a full stack builds, the bblayers.conf.sample file is
updated to include all layers, and activate them if and only if they are
present in the workspace.

i.e. checks are added for the presence of all layers, and activate them
if present.

This remove the need to have different bblayers.conf.sample for different
layer builds. Reducing duplication and fixing path naming convention errors.